refactor: consolidate AddButtons & InsertCellMenu
Keeping it DRY
