+++ ui/src/notebooks/components/header/index.tsx
@@ -15,7 +15,7 @@ const Header: FC = () => (
     <Page.ControlBar fullWidth={FULL_WIDTH}>
       <Page.ControlBarLeft>
         <h3 className="notebook--add-cell-label">Add Cell:</h3>
-        <AddButtons />
+        <AddButtons eventName="Notebook Add Button Clicked" />
       </Page.ControlBarLeft>
       <Page.ControlBarRight>
         <MiniMapToggle />
